backend/pyproject.toml
[project]
name = "trade-journal-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "tzdata>=2025.2",
    "yfinance>=0.2.36",
    "anthropic>=0.39.0",
    "scipy>=1.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app"]
